Got rid of global variables, replaced a bunch of selectors to one single selector, separated api methods from script that is using them
